Revert "feat(front): add more MCP servers around voice (#17138)"
This reverts commit 3bac9c87cb9e0b2c6ea08511c53c34ae8472228d.
