+++ front/lib/actions/mcp_internal_actions/constants.ts
@@ -75,7 +75,6 @@ export const AVAILABLE_INTERNAL_MCP_SERVER_NAMES = [
   "conversation_files",
   "data_sources_file_system",
   DATA_WAREHOUSE_SERVER_NAME,
-  "common_utilities",
   "deep_dive",
   "extract_data",
   "file_generation",
@@ -87,6 +86,7 @@ export const AVAILABLE_INTERNAL_MCP_SERVER_NAMES = [
   "google_sheets",
   "hubspot",
   "image_generation",
+  "elevenlabs",
   "include_data",
   "interactive_content",
   "slideshow",
@@ -108,8 +108,6 @@ export const AVAILABLE_INTERNAL_MCP_SERVER_NAMES = [
   "salesforce",
   "slack",
   "slack_bot",
-  "sound_studio",
-  "speech_generator",
   "toolsets",
   "web_search_&_browse",
   SEARCH_SERVER_NAME,
@@ -1046,7 +1044,7 @@ The directive should be used to display a clickable version of the agent name in
       instructions: null,
     },
   },
-  speech_generator: {
+  elevenlabs: {
     id: 34,
     availability: "manual",
     allowMultipleInstances: false,
@@ -1056,14 +1054,14 @@ The directive should be used to display a clickable version of the agent name in
     isPreview: false,
     tools_stakes: {
       text_to_speech: "low",
-      text_to_discussion: "low",
+      generate_music: "low",
     },
     tools_retry_policies: { default: "retry_on_interrupt" },
     timeoutMs: undefined,
     serverInfo: {
-      name: "speech_generator",
+      name: "elevenlabs",
       version: "1.0.0",
-      description: "Turn written text into spoken audio or dialog",
+      description: "Generate speech audio and music with ElevenLabs.",
       authorization: null,
       icon: "ActionMegaphoneIcon",
       documentationUrl: null,
@@ -1127,30 +1125,6 @@ The directive should be used to display a clickable version of the agent name in
       },
       icon: "MicrosoftTeamsLogo",
       documentationUrl: "https://docs.dust.tt/docs/microsoft-teams-tool-setup",
-      instructions: null,
-    },
-  },
-  sound_studio: {
-    id: 37,
-    availability: "manual",
-    allowMultipleInstances: false,
-    isRestricted: ({ featureFlags }) => {
-      return !featureFlags.includes("elevenlabs_tool");
-    },
-    isPreview: false,
-    tools_stakes: {
-      generate_music: "low",
-      generate_sound_effects: "low",
-    },
-    tools_retry_policies: { default: "retry_on_interrupt" },
-    timeoutMs: undefined,
-    serverInfo: {
-      name: "sound_studio",
-      version: "1.0.0",
-      description: "Create music tracks and sound effects",
-      authorization: null,
-      icon: "ActionMegaphoneIcon",
-      documentationUrl: null,
       instructions: null,
     },
   },
